BE-323: Move Epoch to global module
