+++ app/app.go
@@ -152,7 +152,7 @@ func NewSgnApp(logger log.Logger, db dbm.DB) *sgnApp {
 	authSubspace := app.paramsKeeper.Subspace(auth.DefaultParamspace)
 	bankSupspace := app.paramsKeeper.Subspace(bank.DefaultParamspace)
 	stakingSubspace := app.paramsKeeper.Subspace(staking.DefaultParamspace)
-	subscribeSubspace := app.paramsKeeper.Subspace(subscribe.DefaultParamspace)
+	globalSubspace := app.paramsKeeper.Subspace(global.DefaultParamspace)
 
 	// The AccountKeeper handles address -> account lookups
 	app.accountKeeper = auth.NewAccountKeeper(
@@ -198,6 +198,7 @@ func NewSgnApp(logger log.Logger, db dbm.DB) *sgnApp {
 		app.keyGlobal,
 		app.cdc,
 		ethClient,
+		globalSubspace,
 	)
 
 	app.validatorKeeper = validator.NewKeeper(
@@ -215,7 +216,6 @@ func NewSgnApp(logger log.Logger, db dbm.DB) *sgnApp {
 		ethClient,
 		app.globalKeeper,
 		app.validatorKeeper,
-		subscribeSubspace,
 	)
 
 	app.mm = module.NewManager(
