+++ x/global/types/key.go
@@ -1,5 +1,7 @@
 package types
 
+import sdk "github.com/cosmos/cosmos-sdk/types"
+
 const (
 	// module name
 	ModuleName = "global"
@@ -10,4 +12,17 @@ const (
 
 var (
 	LatestBlockKey = []byte{0x01} // Key for lastest block
+
+	EpochKeyPrefix = []byte{0x21} // Key prefix for epoch
+	LatestEpochKey = []byte{0x22} // Key for latest epoch
 )
+
+// get epoch key from epochId
+func GetEpochKey(epochId sdk.Int) []byte {
+	return append(EpochKeyPrefix, epochId.BigInt().Bytes()...)
+}
+
+// get latest epoch key
+func GetLatestEpochKey() []byte {
+	return LatestEpochKey
+}
