+++ x/subscribe/alias.go
@@ -11,8 +11,6 @@ const (
 	StoreKey          = types.StoreKey
 	QuerySubscription = types.QuerySubscription
 	QueryRequest      = types.QueryRequest
-	QueryEpoch        = types.QueryEpoch
-	QueryParameters   = types.QueryParameters
 )
 
 var (
@@ -23,17 +21,12 @@ var (
 	RegisterCodec              = types.RegisterCodec
 	NewSubscription            = types.NewSubscription
 	NewRequest                 = types.NewRequest
-	NewEpoch                   = types.NewEpoch
 	NewQuerySubscriptionParams = types.NewQuerySubscriptionParams
 	NewQueryRequestParams      = types.NewQueryRequestParams
-	NewQueryEpochParams        = types.NewQueryEpochParams
 	GetSubscriptionKey         = types.GetSubscriptionKey
 	GetRequestKey              = types.GetRequestKey
-	GetEpochKey                = types.GetEpochKey
-	GetLatestEpochKey          = types.GetLatestEpochKey
 	SubscriptionKeyPrefix      = types.SubscriptionKeyPrefix
 	CLIQueryRequest            = cli.QueryRequest
-	DefaultParams              = types.DefaultParams
 )
 
 type (
@@ -42,9 +35,6 @@ type (
 	MsgGuardProof           = types.MsgGuardProof
 	Subscription            = types.Subscription
 	Request                 = types.Request
-	Epoch                   = types.Epoch
-	Params                  = types.Params
 	QuerySubscriptionParams = types.QuerySubscriptionParams
 	QueryRequestParams      = types.QueryRequestParams
-	QueryEpochParams        = types.QueryEpochParams
 )
