+++ x/subscribe/client/cli/query.go
@@ -9,11 +9,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/codec"
 	ethcommon "github.com/ethereum/go-ethereum/common"
 	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
-)
-
-const (
-	flagEpochId = "epochId"
 )
 
 func GetQueryCmd(storeKey string, cdc *codec.Codec) *cobra.Command {
@@ -27,7 +22,6 @@ func GetQueryCmd(storeKey string, cdc *codec.Codec) *cobra.Command {
 	subscribeQueryCmd.AddCommand(client.GetCommands(
 		GetCmdSubscription(storeKey, cdc),
 		GetCmdRequest(storeKey, cdc),
-		GetCmdQueryParams(storeKey, cdc),
 	)...)
 	return subscribeQueryCmd
 }
@@ -102,57 +96,4 @@ func QueryRequest(cdc *codec.Codec, cliCtx context.CLIContext, queryRoute string
 
 	cdc.MustUnmarshalJSON(res, &request)
 	return
-}
-
-// GetCmdEpoch queries request info
-func GetCmdEpoch(queryRoute string, cdc *codec.Codec) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "epoch",
-		Short: "query epoch info by epochId",
-		Args:  cobra.NoArgs,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			epochId := viper.GetInt64(flagEpochId)
-			data, err := cdc.MarshalJSON(types.NewQueryEpochParams(epochId))
-			if err != nil {
-				return err
-			}
-
-			cliCtx := context.NewCLIContext().WithCodec(cdc)
-			route := fmt.Sprintf("custom/%s/%s", queryRoute, types.QueryEpoch)
-			bz, _, err := cliCtx.QueryWithData(route, data)
-			if err != nil {
-				return err
-			}
-
-			var epoch types.Epoch
-			cdc.MustUnmarshalJSON(bz, &epoch)
-			return cliCtx.PrintOutput(epoch)
-		},
-	}
-
-	cmd.Flags().Int64(flagEpochId, 0, "Epoch id")
-
-	return cmd
-}
-
-// GetCmdQueryParams implements the params query command.
-func GetCmdQueryParams(queryRoute string, cdc *codec.Codec) *cobra.Command {
-	return &cobra.Command{
-		Use:   "params",
-		Args:  cobra.NoArgs,
-		Short: "Query the current subscribe parameters information",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			cliCtx := context.NewCLIContext().WithCodec(cdc)
-
-			route := fmt.Sprintf("custom/%s/%s", queryRoute, types.QueryParameters)
-			bz, _, err := cliCtx.QueryWithData(route, nil)
-			if err != nil {
-				return err
-			}
-
-			var params types.Params
-			cdc.MustUnmarshalJSON(bz, &params)
-			return cliCtx.PrintOutput(params)
-		},
-	}
 }
