+++ x/subscribe/keeper.go
@@ -6,7 +6,6 @@ import (
 	"github.com/celer-network/sgn/x/validator"
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/params"
 )
 
 // Keeper maintains the link to data storage and exposes getter/setter methods for the various parts of the state machine
@@ -16,19 +15,17 @@ type Keeper struct {
 	ethClient       *mainchain.EthClient
 	globalKeeper    global.Keeper
 	validatorKeeper validator.Keeper
-	paramstore      params.Subspace
 }
 
 // NewKeeper creates new instances of the subscribe Keeper
 func NewKeeper(storeKey sdk.StoreKey, cdc *codec.Codec, ethClient *mainchain.EthClient,
-	globalKeeper global.Keeper, validatorKeeper validator.Keeper, paramstore params.Subspace) Keeper {
+	globalKeeper global.Keeper, validatorKeeper validator.Keeper) Keeper {
 	return Keeper{
 		storeKey:        storeKey,
 		cdc:             cdc,
 		ethClient:       ethClient,
 		globalKeeper:    globalKeeper,
 		validatorKeeper: validatorKeeper,
-		paramstore:      paramstore.WithKeyTable(ParamKeyTable()),
 	}
 }
 
@@ -85,45 +82,4 @@ func (k Keeper) GetRequest(ctx sdk.Context, channelId []byte) (Request, bool) {
 func (k Keeper) SetRequest(ctx sdk.Context, channelId []byte, request Request) {
 	store := ctx.KVStore(k.storeKey)
 	store.Set(GetRequestKey(channelId), k.cdc.MustMarshalBinaryBare(request))
-}
-
-// Gets the entire Epoch metadata for a epochId
-func (k Keeper) GetEpoch(ctx sdk.Context, epochId sdk.Int) (epoch Epoch, found bool) {
-	store := ctx.KVStore(k.storeKey)
-
-	if !store.Has(GetEpochKey(epochId)) {
-		return epoch, false
-	}
-
-	value := store.Get(GetEpochKey(epochId))
-	k.cdc.MustUnmarshalBinaryBare(value, &epoch)
-	return epoch, true
-}
-
-// Sets the entire Epoch metadata for a epochId
-func (k Keeper) SetEpoch(ctx sdk.Context, epoch Epoch) {
-	store := ctx.KVStore(k.storeKey)
-	store.Set(GetEpochKey(epoch.Id), k.cdc.MustMarshalBinaryBare(epoch))
-}
-
-// Gets the entire latest Epoch metadata
-func (k Keeper) GetLatestEpoch(ctx sdk.Context) (epoch Epoch) {
-	store := ctx.KVStore(k.storeKey)
-
-	if !store.Has(GetLatestEpochKey()) {
-		epoch = NewEpoch(sdk.NewInt(1), ctx.BlockTime().Unix())
-		k.SetLatestEpoch(ctx, epoch)
-		return
-	}
-
-	value := store.Get(GetLatestEpochKey())
-	k.cdc.MustUnmarshalBinaryBare(value, &epoch)
-	return
-}
-
-// Sets the entire LatestEpoch metadata
-func (k Keeper) SetLatestEpoch(ctx sdk.Context, epoch Epoch) {
-	store := ctx.KVStore(k.storeKey)
-	store.Set(GetLatestEpochKey(), k.cdc.MustMarshalBinaryBare(epoch))
-	k.SetEpoch(ctx, epoch)
 }
