+++ x/subscribe/querier.go
@@ -3,7 +3,6 @@ package subscribe
 import (
 	"fmt"
 
-	"github.com/celer-network/sgn/x/subscribe/types"
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	abci "github.com/tendermint/tendermint/abci/types"
@@ -17,10 +16,6 @@ func NewQuerier(keeper Keeper) sdk.Querier {
 			return querySubscription(ctx, req, keeper)
 		case QueryRequest:
 			return queryRequest(ctx, req, keeper)
-		case QueryEpoch:
-			return queryEpoch(ctx, req, keeper)
-		case QueryParameters:
-			return queryParameters(ctx, keeper)
 		default:
 			return nil, sdk.ErrUnknownRequest("Unknown subscribe query endpoint")
 		}
@@ -63,44 +58,6 @@ func queryRequest(ctx sdk.Context, req abci.RequestQuery, keeper Keeper) ([]byte
 	if err != nil {
 		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("Could not marshal result to JSON", err.Error()))
 
-	}
-
-	return res, nil
-}
-
-func queryEpoch(ctx sdk.Context, req abci.RequestQuery, keeper Keeper) ([]byte, sdk.Error) {
-	var params QueryEpochParams
-	err := ModuleCdc.UnmarshalJSON(req.Data, &params)
-	if err != nil {
-		return nil, sdk.ErrInternal(fmt.Sprintf("Failed to parse params: %s", err))
-	}
-
-	var epoch Epoch
-
-	if params.EpochId <= 0 {
-		epoch = keeper.GetLatestEpoch(ctx)
-	} else {
-		epoch, _ = keeper.GetEpoch(ctx, sdk.NewInt(params.EpochId))
-		if epoch.Id.IsZero() {
-			return nil, sdk.ErrInternal("Could not find corresponding epoch")
-		}
-	}
-
-	res, err := codec.MarshalJSONIndent(keeper.cdc, epoch)
-	if err != nil {
-		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("Could not marshal result to JSON", err.Error()))
-
-	}
-
-	return res, nil
-}
-
-func queryParameters(ctx sdk.Context, k Keeper) ([]byte, sdk.Error) {
-	params := k.GetParams(ctx)
-
-	res, err := codec.MarshalJSONIndent(types.ModuleCdc, params)
-	if err != nil {
-		return nil, sdk.ErrInternal(sdk.AppendMsgToErr("could not marshal result to JSON", err.Error()))
 	}
 
 	return res, nil
