+++ x/subscribe/types/key.go
@@ -1,9 +1,5 @@
 package types
 
-import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
-)
-
 const (
 	// module name
 	ModuleName = "subscribe"
@@ -15,9 +11,6 @@ const (
 var (
 	SubscriptionKeyPrefix = []byte{0x01} // Key prefix for subscription
 	RequestKeyPrefix      = []byte{0x02} // Key prefix for request
-
-	EpochKeyPrefix = []byte{0x31} // Key prefix for epoch
-	LatestEpochKey = []byte{0x32} // Key for latest epoch
 )
 
 // get guardian key from eth address
@@ -28,14 +21,4 @@ func GetSubscriptionKey(ethAddress string) []byte {
 // get request key from channelID
 func GetRequestKey(channelId []byte) []byte {
 	return append(RequestKeyPrefix, channelId...)
-}
-
-// get epoch key from epochId
-func GetEpochKey(epochId sdk.Int) []byte {
-	return append(EpochKeyPrefix, epochId.BigInt().Bytes()...)
-}
-
-// get latest epoch key
-func GetLatestEpochKey() []byte {
-	return LatestEpochKey
 }
